+++ resources/views/master/librarian_master.blade.php
@@ -105,7 +105,7 @@
                         <span class="nav-link-title"> Borrows </span>
                       </a>
                     </li>
-                    <li class="nav-item dropdown">
+                    <li class="nav-item dropdown {{ request()->routeIs('librarian.category.*') || request()->routeIs('librarian.manageBooks.*') ? 'active' : '' }}">
                       <a class="nav-link dropdown-toggle" href="#navbar-form" data-bs-toggle="dropdown" data-bs-auto-close="outside" role="button" aria-expanded="false">
                         <span class="nav-link-icon d-md-none d-lg-inline-block"><!-- Download SVG icon from http://tabler.io/icons/icon/checkbox -->
                           <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="icon icon-tabler icons-tabler-outline icon-tabler-api-book">
@@ -127,7 +127,7 @@
                       </a>
                       <div class="dropdown-menu">
                         <a class="dropdown-item" href="{{ route('librarian.category.index') }}"> Categories </a>
-                        <a class="dropdown-item" href="{{ route('librarian.dashboard') }}"> Manage Books </a>
+                        <a class="dropdown-item" href="{{ route('librarian.manageBooks.index') }}"> Manage Books </a>
                       </div>
                     </li>
                     <li class="nav-item {{ request()->routeIs('librarian.author.*') ? 'active' : '' }}">